assets/pallet-verification-agent/requirements.txt
litellm==1.83.0,!=1.82.7,!=1.82.8
langchain==1.2.15
langchain-core==1.3.0
langchain-litellm==0.3.5
langgraph==1.1.9
a2a-sdk[all]==0.3.22
uvicorn==0.40.0
httpx==0.28.1
python-dotenv==1.2.1
click==8.3.1
sap-cloud-sdk @ https://common.repositories.cloud.sap/artifactory/api/pypi/application-foundation-sdk-python/sap-cloud-sdk/0.11.3.dev20260423+contribs.sapphire.internal.testing.0a2505a/sap_cloud_sdk-0.11.3.dev20260423+contribs.sapphire.internal.testing.0a2505a-py3-none-any.whl
mcp>=1.0.0
